+++ CHANGES.md
@@ -44,9 +44,11 @@ Beta Releases
 * `GeometryVisualizer` now creates geometry asynchronously to prevent locking up the browser.
 * `Clock.canAnimate` was added to prevent time from advancing, even while the clock is animating.
 * `Viewer` now prevents time from advancing if asynchronous geometry is being processed in order to avoid showing an incomplete picture.  This can be disabled via the `Viewer.allowDataSourcesToSuspendAnimation` settings.
+* Added `Model.minimumPixelSize` property so models remain visible when the viewer zooms out.
 * Added ability to modify glTF material parameters using `Model.getMaterial`, `ModelMaterial`, and 'ModelMesh.material'.
-* Added `sizeInMeters`, `asynchronous`, and `ready` properties to `Model`.
+* Added `asynchronous` and `ready` properties to `Model`.
 * Added `Cartesian4.fromColor` and `Color.fromCartesian4`.
+* Added `getScale` and `getMaximumScale` to `Matrix2`, `Matrix3`, and `Matrix4`.
 
 ### b27 - 2014-04-01